Add mode field to Sprite (#563)
* Added sprite_hide and sprite_object_hide functions

* Changed calls to obj_unhide to sprite_unhide

* Fix embarrasing errors

* clang-format...?

* Imported renames from #557

* Removed last uses of obj_hide in round

* Added documentation

* Changed Sprite.idx from int to s16

* Added proper NULL-check macro usage in sprite.c

+++ include/card.h
@@ -75,8 +75,8 @@ u8 card_get_value(Card* card);
 // CardObject methods
 CardObject* card_object_new(Card* card);
 void card_object_destroy(CardObject** card_object);
-void card_object_set_sprite(CardObject* card_object, int layer);
-void card_object_set_sprite_face_down(CardObject* card_object, enum DeckType deck, int layer);
+void card_object_set_sprite(CardObject* card_object, s16 layer);
+void card_object_set_sprite_face_down(CardObject* card_object, enum DeckType deck, s16 layer);
 void card_object_shake(CardObject* card_object, mm_word sound_id);
 
 void card_object_set_selected(CardObject* card_object, bool selected);
